resolution d'autres problemes
